Remove duplicated diagnostics
Currently, ts-jest is checking the type errors on the same source file
at least twice:
1. Preprocess the test files, diagnostics will happen on all test files
and its dependencies
2. Preprocess the imported source files, when import a file in the test
code or source code, it is always preprocessed then it will execute
the run-diagnostics again. (this one should be removed)

Add the logic to check whether the preprocessed file match testRegex or
testMatch, if so, run diagnostic, otherwise skip this process, because
it has been diagnosed during preprocessing the test file.